Day5/Study_Assistant/requirements.txt

# streamlit
# PyPDF2
# openai

streamlit
PyPDF2
google-generativeai